requirements.txt
distlib==0.3.9
filelock==3.18.0
pbr==6.1.1
platformdirs==4.3.8
stevedore==5.4.1
virtualenv==20.31.2
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.1
pandas
psycopg2-binary
python-dotenv
